move common rpc class to common api project
