+++ CommonAPI/CommonAPI.csproj
@@ -31,6 +31,7 @@
 		<PackageReference Include="Azure.Security.KeyVault.Secrets" Version="[4.5.0,5.0)" />
 		<PackageReference Include="BrassLoon.Interface.Account" Version="4.0.*" />
 		<PackageReference Include="BrassLoon.Interface.Log" Version="4.0.*" />
+		<PackageReference Include="Grpc.AspNetCore" Version="2.57.*" />
 		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="[6.0.14,7.0)" />
 		<PackageReference Include="Polly" Version="[7.2.3,8.0)" />
 		<PackageReference Include="Polly.Caching.Memory" Version="[3.0.2,4.0)" />
