+++ CommonAPI/DomainAcountAccessVerifier.cs
@@ -1,30 +1,25 @@
 ﻿using BrassLoon.Interface.Account;
 using BrassLoon.Interface.Account.Models;
-using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace LogRPC
+namespace BrassLoon.CommonAPI
 {
     public class DomainAcountAccessVerifier : IDomainAcountAccessVerifier
-    {   
-        private readonly SettingsFactory _settingsFactory;
+    {
         private readonly IDomainService _domainService;
 
-        public DomainAcountAccessVerifier(
-            SettingsFactory settingsFactory,
-            IDomainService domainService)
+        public DomainAcountAccessVerifier(IDomainService domainService)
         {
-            _settingsFactory = settingsFactory;
             _domainService = domainService;
         }
 
-        public async Task<bool> HasAccess(Settings settings, Guid domainId, string accessToken)
+        public async Task<bool> HasAccess(ISettings settings, Guid domainId, string accessToken)
         {
-            AccountDomain accountDomain = await GetDomain(domainId, settings, accessToken, _domainService);
+            AccountDomain accountDomain = await GetDomain(_domainService, domainId, settings);
             return accountDomain != null && !accountDomain.Account.Locked && VerifyDomainAccount(accountDomain, accessToken);
         }
 
@@ -35,18 +30,16 @@ namespace LogRPC
 
         private bool UserCanAccessAccount(Guid accountId, string accessToken)
         {
-            JwtSecurityToken jwtSecurityToken = new JwtSecurityToken(accessToken);            
-            string[] accountIds = Regex.Split(jwtSecurityToken.Claims.First(c => c.Type == "accounts").Value, @"\s+", RegexOptions.IgnoreCase);
+            JwtSecurityToken jwtSecurityToken = new JwtSecurityToken(accessToken);
+            string[] accountIds = Regex.Split(jwtSecurityToken.Claims.First(c => c.Type == "accounts").Value, @"\s+", RegexOptions.IgnoreCase, TimeSpan.FromMilliseconds(200));
             return accountIds.Where(id => !string.IsNullOrEmpty(id)).Any(id => Guid.Parse(id).Equals(accountId));
         }
 
-        private async Task<AccountDomain> GetDomain(Guid domainId, Settings settings, string accessToken, IDomainService domainService)
+        private static async Task<AccountDomain> GetDomain(IDomainService domainService, Guid domainId, ISettings settings)
         {
             try
             {
-                return await domainService.GetAccountDomain(
-                    _settingsFactory.CreateAccount(settings, accessToken),
-                    domainId);
+                return await domainService.GetAccountDomain(settings, domainId);
             }
             catch (BrassLoon.RestClient.Exceptions.RequestError ex)
             {
