+++ CommonAPI/IDomainAcountAccessVerifier.cs
@@ -1,0 +1,10 @@
+﻿using BrassLoon.Interface.Account;
+using System.Threading.Tasks;
+
+namespace BrassLoon.CommonAPI
+{
+    public interface IDomainAcountAccessVerifier
+    {
+        Task<bool> HasAccess(ISettings settings, Guid domainId, string accessToken);
+    }
+}
