--- Log/LogRPC/IDomainAcountAccessVerifier.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace LogRPC
-{
-    public interface IDomainAcountAccessVerifier
-    {
-        Task<bool> HasAccess(Settings settings, Guid domainId, string accessToken);
-    }
-}
